pyproject.toml
[project]
name = "daily-real"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.12.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "testcontainers>=4.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "testcontainers>=4.10.0",
]
